new fix

+++ dilab_server_side.js
@@ -921,10 +921,13 @@ app.post("/Dilab/:action", upload.array("files"), (req,res,err) => {
                                 });
                             }
                             if (req.files) {
-                                for (var i=0;i<req.files.length;i++)
-                                    fs.unlinkSync(req.files[i].path,()=>{return;});
-                                    i--;
+                                for (var i=0;i<req.files.length;i++) {
+                                    if (req.files[i]!=filePath3 && req.files[i]!=filePath1 && req.files[i]!=filePath2) {
+                                        fs.unlinkSync(req.files[i].path,()=>{return;});
+                                        i--;                                      
+                                    } 
                                 }
+                            }
                         });
                     }
                 });
